step1, 1주차 기능 구현
