--- src/main/java/Status.java
@@ -1,3 +1,0 @@
-public enum Status { //
-    DONE, YET
-}
