+++ src/main/java/org/homework/TodoItem.java
@@ -1,0 +1,13 @@
+package org.homework;
+
+public class TodoItem {
+    private String title;
+
+    public TodoItem(String title) {
+        this.title = title;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+}
